+++ packages/web3-validator/src/validator.ts
@@ -16,53 +16,93 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 import { Web3ValidationErrorObject } from 'web3-types';
 
-import { toHex, utf8ToBytes } from 'ethereum-cryptography/utils.js';
-import { blake2b } from 'ethereum-cryptography/blake2b.js';
-import validator from 'is-my-json-valid';
-import formats from './formats.js';
+import { z, ZodType, ZodIssue, ZodIssueCode, ZodTypeAny } from 'zod';
+
+import { RawCreateParams } from 'zod/lib/types';
 import { Web3ValidatorError } from './errors.js';
-import { Validate, Json, Schema, RawValidationError } from './types.js';
+import { Json, JsonSchema } from './types.js';
+import formats from './formats';
+
+const convertToZod = (schema: JsonSchema): ZodType => {
+	if ((!schema?.type || schema?.type === 'object') && schema?.properties) {
+		const obj: { [key: string]: ZodType } = {};
+		for (const name of Object.keys(schema.properties)) {
+			const zItem = convertToZod(schema.properties[name]);
+			if (zItem) {
+				obj[name] = zItem;
+			}
+		}
+
+		if (Array.isArray(schema.required)) {
+			return z
+				.object(obj)
+				.partial()
+				.required(schema.required.reduce((acc, v: string) => ({ ...acc, [v]: true }), {}));
+		}
+		return z.object(obj).partial();
+	}
+
+	if (schema?.type === 'array' && schema?.items) {
+		if (Array.isArray(schema.items) && schema.items.length > 0) {
+			const arr: Partial<[ZodTypeAny, ...ZodTypeAny[]]> = [];
+			for (const item of schema.items) {
+				const zItem = convertToZod(item);
+				if (zItem) {
+					arr.push(zItem);
+				}
+			}
+			return z.tuple(arr as [ZodTypeAny, ...ZodTypeAny[]]);
+		}
+		return z.array(convertToZod(schema.items as JsonSchema));
+	}
+
+	if (schema.oneOf && Array.isArray(schema.oneOf)) {
+		return z.union(
+			schema.oneOf.map(oneOfSchema => convertToZod(oneOfSchema)) as [
+				ZodTypeAny,
+				ZodTypeAny,
+				...ZodTypeAny[],
+			],
+		);
+	}
+
+	if (schema?.format) {
+		return z.any().refine(formats[schema.format], (value: unknown) => ({
+			params: { value, format: schema.format },
+		}));
+	}
+
+	if (
+		schema?.type &&
+		schema?.type !== 'object' &&
+		typeof (z as unknown as { [key: string]: (params?: RawCreateParams) => ZodType })[
+			String(schema.type)
+		] === 'function'
+	) {
+		return (z as unknown as { [key: string]: (params?: RawCreateParams) => ZodType })[
+			String(schema.type)
+		]();
+	}
+	return z.object({ data: z.any() }).partial();
+};
 
 export class Validator {
 	// eslint-disable-next-line no-use-before-define
 	private static validatorInstance?: Validator;
+
 	// eslint-disable-next-line no-useless-constructor, @typescript-eslint/no-empty-function
-	private constructor() {}
 	public static factory(): Validator {
 		if (!Validator.validatorInstance) {
 			Validator.validatorInstance = new Validator();
 		}
 		return Validator.validatorInstance;
 	}
-	private readonly _schemas: Map<string, Validate> = new Map();
-	public getSchema(key: string) {
-		return this._schemas.get(key);
-	}
 
-	public addSchema(key: string, schema: Schema) {
-		this._schemas.set(key, this.createValidator(schema));
-	}
-
-	// eslint-disable-next-line  class-methods-use-this
-	private createValidator(schema: Schema): Validate {
-		// eslint-disable-next-line  @typescript-eslint/no-unsafe-call
-		// @ts-expect-error validator params correction
-		return validator(schema, {
-			formats,
-			greedy: true,
-			verbose: true,
-			additionalProperties: false,
-		}) as Validate;
-	}
-
-	public validate(schema: Schema, data: Json, options?: { silent?: boolean }) {
-		const localValidate = this.getOrCreateValidator(schema);
-		if (!localValidate(data)) {
-			const errors = this.convertErrors(
-				localValidate.errors as RawValidationError[],
-				schema,
-				data,
-			);
+	public validate(schema: JsonSchema, data: Json, options?: { silent?: boolean }) {
+		const zod = convertToZod(schema);
+		const result = zod.safeParse(data);
+		if (!result.success) {
+			const errors = this.convertErrors(result.error?.issues ?? []);
 			if (errors) {
 				if (options?.silent) {
 					return errors;
@@ -72,129 +112,57 @@ export class Validator {
 		}
 		return undefined;
 	}
-	private convertErrors(
-		errors: RawValidationError[] | undefined,
-		schema: Schema,
-		data: Json,
-	): Web3ValidationErrorObject[] | undefined {
+	// eslint-disable-next-line class-methods-use-this
+	private convertErrors(errors: ZodIssue[] | undefined): Web3ValidationErrorObject[] | undefined {
 		if (errors && Array.isArray(errors) && errors.length > 0) {
-			return errors.map((error: RawValidationError) => {
+			return errors.map((error: ZodIssue) => {
 				let message;
 				let keyword;
 				let params;
 				let schemaPath;
 
-				schemaPath = Array.isArray(error.schemaPath)
-					? error.schemaPath.slice(1).join('/')
-					: '';
+				schemaPath = error.path.join('/');
 
-				const { field } = error;
-				const _instancePath =
-					schemaPath ||
-					// eslint-disable-next-line no-useless-escape
-					(field?.length >= 4 ? `${field.slice(4).replace(/\"|\[|\]/g, '')}` : '/');
+				const field = String(error.path[error.path.length - 1]);
+				const instancePath = error.path.join('/');
+				if (error.code === ZodIssueCode.too_big) {
+					keyword = 'maxItems';
+					schemaPath = `${instancePath}/maxItems`;
+					params = { limit: error.maximum };
+					message = `must NOT have more than ${error.maximum} items`;
+				} else if (error.code === ZodIssueCode.too_small) {
+					keyword = 'minItems';
+					schemaPath = `${instancePath}/minItems`;
+					params = { limit: error.minimum };
+					message = `must NOT have fewer than ${error.minimum} items`;
+				} else if (error.code === ZodIssueCode.custom) {
+					const { value, format } = (error.params ?? {}) as {
+						value: unknown;
+						format: string;
+					};
 
-				const instancePath = _instancePath ? `/${_instancePath}` : '';
-				if (error?.message === 'has less items than allowed') {
-					// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-					const schemaData = this.getObjectValueByPath(schema, schemaPath);
-					// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-					if (schemaData.minItems) {
-						keyword = 'minItems';
-						schemaPath = `${schemaPath}/minItems`;
-						// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access
-						params = { limit: schemaData.minItems };
-						// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/restrict-template-expressions
-						message = `must NOT have fewer than ${schemaData.minItems} items`;
+					if (typeof value === 'undefined') {
+						message = `value at "/${schemaPath}" is required`;
+					} else {
+						message = `value "${
+							// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+							typeof value === 'object' ? JSON.stringify(value) : value
+						}" at "/${schemaPath}" must pass "${format}" validation`;
 					}
-				} else if (error?.message === 'has more items than allowed') {
-					// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-					const schemaData = this.getObjectValueByPath(schema, schemaPath);
-					// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-					if (schemaData.maxItems) {
-						keyword = 'maxItems';
-						schemaPath = `${schemaPath}/maxItems`;
-						// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access
-						params = { limit: schemaData.maxItems };
-						// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/restrict-template-expressions
-						message = `must NOT have more than ${schemaData.maxItems} items`;
-					}
-				} else if (
-					error?.message.startsWith('must be') &&
-					error?.message.endsWith('format')
-				) {
-					const formatName = error?.message.split(' ')[2];
-					if (formatName) {
-						message = `must pass "${formatName}" validation`;
-					}
+
+					params = { value };
 				}
-				// eslint-disable-next-line  @typescript-eslint/no-unsafe-assignment
-				const dataValue = this.getObjectValueByPath(data as object, instancePath);
+
 				return {
-					keyword: keyword ?? error.field,
-					instancePath,
-					schemaPath: `#${schemaPath}`,
-					// eslint-disable-next-line  @typescript-eslint/no-unsafe-assignment
-					params: params ?? { value: dataValue },
+					keyword: keyword ?? field,
+					instancePath: instancePath ? `/${instancePath}` : '',
+					schemaPath: schemaPath ? `#${schemaPath}` : '#',
+					// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+					params: params ?? { value: error.message },
 					message: message ?? error.message,
 				} as Web3ValidationErrorObject;
 			});
 		}
 		return undefined;
-	}
-
-	public getOrCreateValidator(schema: Schema): Validate {
-		const key = Validator.getKey(schema);
-		let _validator = this.getSchema(key);
-		if (!_validator) {
-			this.addSchema(key, schema);
-			_validator = this.getSchema(key);
-		}
-		return _validator!;
-	}
-
-	public static getKey(schema: Schema) {
-		return toHex(blake2b(utf8ToBytes(JSON.stringify(schema))));
-	}
-	private getObjectValueByPath(obj: object, pointer: string, objpath?: object[]) {
-		try {
-			if (typeof obj !== 'object') throw new Error('Invalid input object');
-			if (typeof pointer !== 'string') throw new Error('Invalid JSON pointer');
-			const parts = pointer.split('/');
-			if (!['', '#'].includes(parts.shift() as string)) {
-				throw new Error('Invalid JSON pointer');
-			}
-			if (parts.length === 0) return obj;
-
-			let curr: any = obj;
-			for (const part of parts) {
-				if (typeof part !== 'string') throw new Error('Invalid JSON pointer');
-				// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-argument
-				if (objpath) objpath.push(curr); // does not include target itself, but includes head
-				const prop = this.untilde(part);
-				if (typeof curr !== 'object') return undefined;
-				if (!Object.prototype.hasOwnProperty.call(curr, prop)) return undefined;
-				// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access
-				curr = curr[prop];
-			}
-			// eslint-disable-next-line @typescript-eslint/no-unsafe-return
-			return curr;
-		} catch (e) {
-			return '';
-		}
-	}
-	// eslint-disable-next-line class-methods-use-this
-	private untilde(string: string) {
-		if (!string.includes('~')) return string;
-		return string.replace(/~[01]/g, match => {
-			switch (match) {
-				case '~1':
-					return '/';
-				case '~0':
-					return '~';
-				default:
-					throw new Error('Unreachable');
-			}
-		});
 	}
 }